Some changes to make MX::Types::Combine detect errors earlier
We check that all the packages we are combining actually have a ->type_names
method.

Type libs are loaded when the provided type libs are set, rather than every
time import is called.

Generate the list of exported types when the provided type libs are set,
rather than every time import is called.