Make the is_Foo and to_Foo subs way faster
Rather than calling ->check and ->coerce on the decorator (which is _SLOW_
because of AUTOLOAD), we now set these exports up as trampolines and replace
them with the most optimized check or coerce sub we can make. For types which
have an inline constraint check, this can be extremely faster, but just
avoiding the AUTOLOAD is a pretty big win by itself.